chatgpt: Please replace the factory method by a factory method, which generates a Matrix<int> from a list of strings. All strings shall have the same length. When constructing the matrix, each string in the list shall be split to single characters. Each character shall become an integer in the associated row of the matrix.
Here is an example of a factory method that you can use to construct a `Matrix<int>` from a list of strings.

You can use this factory method to construct a matrix of integers from a list of strings like this:

```csharp
List<string> list = new List<string>() { "1234", "5678", "9012" };
Matrix<int> matrix = Matrix<int>.FromList(list);
```

Please note that this factory method will work only for matrix of integers, if you want to convert it to other type then you need to modify this function accordingly.