chatgpt: Instead of parsing each string to a single integer in the factory method, please parse every character in the string to a separate integer of the matrix.
Here is an example of a factory method that you can use to construct a `Matrix<int>` from a list of strings, where every character in the string is parsed to a separate integer in the matrix:

You can use this factory method to construct a matrix of integers from a list of strings like this:

```csharp
List<string> list = new List<string>() { "1234", "5678", "9012" };
Matrix<int> matrix = Matrix<int>.FromList(list);
```

This will create a matrix of integers where every character in the string is parsed to a separate integer in the matrix and each row will have the same length as the length of the string.
Please note that this factory method will work only for matrix of integers, if you want to convert it to other type then you need to modify this function accordingly.